+++ state/stateManager.js
@@ -7,9 +7,9 @@ export const createStates = (gridSize) => {
     }; 
     sealObject(animationState);
     const concentrationState = {
-        concentrationField: new Float32Array(gridSize).fill(0.001),
-        sources: new Float32Array(gridSize).fill(0),
-        sinks:new Float32Array(gridSize).fill(0),
+        concentrationField: new Float64Array(gridSize).fill(0.001),
+        sources: new Float64Array(gridSize).fill(0),
+        sinks:new Float64Array(gridSize).fill(0),
         visible:null,
     }
     sealObject(concentrationState);
